Add an audio envelope class
Envelopes each channel's initial audio samples permitting the maximum
allowed volume to rapidly expand from zero to max-amplitude within a
period as short as 15ms.

The goal is to catch misbehaving "audio" samples such as binary data,
digital machine noise, or brick-wall DC-offsets from causing clicks or
pops within the split-second of initial audio. Its secondary goal is to
not  (or minimally) impact the volume of "organic" or natural waveforms,
which (if recorded properly) start at zero and ramp up and down along
sine/cosine curves.

This was implemented in a very efficient manner:

 -  The envelope parameters are pre-computed once when the channel is
    configured

 -  The envelope itself consists of a minimal set of integer
    comparisons. Once the envelope has been fully expanded or expired,
    it's replaced with a no-op function so completely gets out of the
    way.

An example can be seen in PC speaker audio. This channel wants to
start straight into playing a square wave, which the envelope captures
and limits the first ~hundred samples (making the pop inaudible) - and
by the time the second square wave, the envelope is out of the picture
with the peak volume permitted.

When compared to well behaved natural audio, the envelope is available
but expires unused. We see that well behaved audio typically faded in
on a far greater time scale (versus the envelopes 15ms), and the volumes
never approach the peak available.  In this case, the envelope isn't
used, and that's exactly what we want. The envelope is the worst-case
fallback.