Change mkdir permissions to 775
By default on Linux, if you create a directory, it is created with
permissions 775 (rwxrwxr-x). But for some reason if you use
DOSBox to create a directory using folder mounts it creates it
with 700 (rwx------)

On the other hand, if you create a file, it creates with the same
permissions as Linux does (664, or rw-rw-r--).

With this patch, DOSBox-staging creates dirctories with the same
permissions as Linux does