Remove unused disasm_tables header
This header was never included anywhere throughout the history of the
project. It was part of the first commit to CVS and never used
afterwards.

Carrying this dead code results in slight inconvenience due to
statistics being skewed and various tools mis-classifying this file
(e.g. GitHub Linguist thinks it's Objective-C file).

If this will be needed (unlikely), the file can be easily found in the
first commit in the history of the project.