Fix copyright attributions in the CD-DA backend
Prior to the creation of the dosbox-staging project, I had ascribed
copyright of the added CD-DA backend source files to The DOSBox Team in
anticipation of it being merged into their project.

This has not come to fruition as The DOSBox Team has not accepted this
work.

To avoid misrepresenting the source of this work (for projects, that
decide to use dosbox-staging as upstream), this PR corrects the
Copyright attribution in these files.

Signed-off-by: Patryk Obara <dreamer.tan@gmail.com>