Use alignment-safe functions in dynrec caches
These changes are designed to remove pointless implementation
differences between dyn_x86 and dynrec core caches.  The responsibility
for handling alignment and endianness is pushed to our mem_host and
mem_unaligned headers (both of which are inline-friendly), so there's no
need for keeping micro-optimizations in this code.

This commit includes two kind of replacements:

1. Simple bool-checks:

- if (!*(Bit16u*)&write_map[addr]) {
+ // see if there's code where we are writing to
+ if (!read_unaligned_uint16(&write_map[addr])) {

In this case, value read from write_map is only checked for being 0, so
there's no need to byteswap it around on big endian machines;
consequently we can get away with read_unaligned_* functions.

2. Marking memory in "invalidation map"

- #if defined(WORDS_BIGENDIAN) || !defined(C_UNALIGNED_MEMORY)
-     host_writew(&invalidation_map[addr],
-                 host_readw(&invalidation_map[addr])+0x101);
- #else
-     (*(Bit16u*)&invalidation_map[addr])+=0x101;
- #endif
+     host_addw(&invalidation_map[addr], 0x0101);

Preserving the byteswapping operation the upstream is using.