Migrate the GUS volume indexes to integers
This allows the math in VolUpdate() to be performed without
casting or roll-over effects using unsigned.

This technically allows 'vol_index_current' to be incremented
into negative territory, which is illegal for an array index
(and might occur if other parts of the logic are bugged, so we
assert a sanity check at the point just before we use it to
lookup the volume scalar from the array.