Use an unordered_map with O(1) lookup to manage the IO handlers
Unordered maps offer the same random lookup time as an array,
as a tradeoff for consuming more memory consumed and longer
insert times versus a std::map.

In our cases the combined added memory comes to 336 bytes, which
is 56-extra bytes for each of the six handlers.  The added insert
time does not affect us, because our inserts take place during
handler installation during the startup phase of dosbox, and stays
constant for the remainder of play.

Note that this doesn't actually reduce the number of handlers
though, because the IO class constructor installs the IO_ReadDefault
and IO_WriteDefault functions into every handler; all IO_MAX slots.

This makes sense when you're using a hardcoded array size, as it
doesn't change the memory used.