Call the IO default handlers directly instead using a lookup
Instead of pre-populating 65k IO handler slots with the default
"dummy" IO_ReadDefault() and IO_WriteDefault() functions, this
commit call these directly if a handler hasn't been installed.

The runtime tradeoff is that we perform an "if-check" instead of
a memory lookup; and the former appears to save a considable amount
of time as the dosbox's startup time is now cut in half.

The other big win is in memory saving: the std::unordered maps
now only contain the handlers that are installed, which we see is
actually not many from the termination diagnostics:

IOBUS: Releasing 136 read and 130 write 8-bit port handlers
IOBUS: Releasing 50 read and 52 write 16-bit port handlers
IOBUS: Releasing 33 read and 35 write 32-bit port handlers
IOBUS: Handlers consumed 14288 total bytes