Bump the seek-table version from 4 to 5
The move to xxHash 0.8 finalizes the algorithm and it's now considered
stable. Its hashes can be relied upon to be identical for all future
versions.

Because of this change, we bump the seek-table identifier which will
cause it to regenerate old mp3 seektables and repopulate them with
new hash values.

Also drops the need for the xxh3.h file, as it's now fully embedded
within xxhash.h.