Cleanup a casting warning in dos drive_iso
This changes the type of the sector from signed to unsigned
(sectors are zero-based and always positive). Likewise, the
drive->readSector(...) function takes in a unsigned sector,
so this is where the warnings are fixed.

Side note:

Sectors should never be negative, so we see this caching code has
a design issue where the cache's position is used to also hold the
state of the cache (a negative value indicates the cache isn't
populated).

Because the maximum sector size in FAT32 is 2^32 (unsigned),
The design decision to use an "int" to store the cached sector
breaks down when sector sizes grow beyond 2^31, which the cache
will not be able to handle.