Simplify drive_local's delete function with actionable checks only
We can either delete the file straight-away or dosbox happened to be
holding the file open, in which case we close it and try again.
Failing those, then the file simply cannot be removed due to OS
security or file-system access controls.

Extra checks to 'fopen' and 'stat' do nothing to alter the state
of either of the above attempts, so these are eliminated.

The original code was also vulnerable to a time-of-check time-of-use
vulnerability (CWE-367), which allows the underly file associated with
the filename to be changed between the check (call to stat) and
the use (call to unlink).