Slightly improve readability of two shell_batch functions
These functions are quite difficult to read, so this commit
attempts to used self-describing variable names.

The DOS_ReadFile() API takes in a uint8_t* to write its
data, however the existing code then operated directly on that
value as if it was a char, producing several implicit sign
conversion warning.

This commit differentiates the "data holder", a uint8_t, versus
the resulting value, which is a char. This fixes two sign
conversion warnings.