Fix regression involving international characters in batch file
International ascii characters fall between 127 and 255, however when
cast into a char these "high ASCII" values wrap to the negative side
of the char.

So this sequence failed to handle those international characters:

val = static_cast<char>(data); // international symbols
                               // now on the negative side ..
if (val > 31)
  ... handle the characters ...

Given we do need to operate on the symbols are char (and store them in
char arrays, such as cmd_write and cmd_read), the solution was to
acknowledge what's happening and permit wrapped high-ASCII in these two
checks.

Fixes #555