Fix a CD-DA roll-over bug when calculating PCM frames
In some cases the numerator in the PCM frame calculation can exceed
a uint32_t causing it to roll-over.  This small change pre-divides
the PCM rate by the Redbook rate, preventing the subsequent
multiplication from ever getting close to limit of a uint32_t.

This fixes a regression when playing Descent 2 with a bin/cue pair:
after starting the first level the CD-DA only plays for 40 seconds.
This is because the game requests playback of track 2, but requests
a duration significantly exceeding the actual track's length (which
is legal).  This commit fixes the PCM frame calculation and thus
the music continues to play into subsequent tracks as expected.

The problem affected only cue/bin pairs (regression was not affecting
audio ripped e.g. as flac or opus).