Extend buildsystem with ability to run unit tests
Using Google Test framework, as it's the best in class.  This solution
is purposefully simple to make it possible to integrate with MSVC
buildsystem in the future.

Provide an example test file with simple tutorial on writing unit tests.

When user enables the feature with --enable-tests, we check if the
system has gtest libs installed: first statically (as some Linux distros
provide gtest only as a static lib), then as shared lib (for other
distros).

Google Test official guide advises to always compile gtest specifically
for your project, to avoid breaking C++ "one definition rule", but so
far testing indicated no problems with this solution for us - perhaps
the situation would be different if we were integrating Google Mock
framework as well (but we don't).

As long as we use autoconf and VS/MSVC, there are considerable issues
with building gtest locally (slower CI, we would need to use Git
submodule, gtest supports only CMake, problematic with extending UT
support to VS builds).