Remove unneeded AC_* macro invocations
AC_CHECK_SIZEOF - these used to be needed to define Bitu* family of
types, but we they are no longer needed (we depend on cstdint instead).
We keep check for pointer size, as the result is used later in the file.

AC_TYPE_SIZE_T - We don't need to check for size_t type; it's guaranteed
to be there, because C++14 mandates C99.

AC_C_CONST - Quote from autoconf manual: "This macro is obsolescent,
as current C compilers support const. New programs need not use this
macro."