Include FluidSynth or its dependency in the package-list scripts
Adds either FluidSynth-2.x itself or glib, its immediate dependency,
to the list of packages to install.  The selection depends on
which of the two (if any) are supported by the repository.

In the event the repo only support glib but is missing FluidSynth 2.x,
the user can then trivially build their own using the contrib builder.