Optionally scale a soundfont's levels by a given percentage
Soundfonts can have wildly different average amplitude levels resulting in
inconsistent output volumes when a user different sound fonts for various
games.

This commit lets users adjust a given soundfont's output level and thereby
allowing them to normalize the output volume, independent of the levels used
(or prefered) for a given game.

User can specify this as an optional second parameter following the soundfont
filename in the "soundfont = " configuration parameter.

Values can range from 1 to 500; or 1% up to 500% of the original levels.
If an invalid value is provided, it defaults to no-scaling or native levels.

For example:

soundfont = /path to/soundfont.sf2       # default, no scaling up or down.
soundfont = /path to/soundfont.sf2  1    # attentuates to 1% of original.
soundfont = /path to/soundfont.sf2  50   # attenuates to 50% of original.
soundfont = /path to/soundfont.sf2  200  # amplifies by 200% of original.
soundfont = /path to/soundfont.sf2  800  # out of bounds, reset to default.
soundfont = /path to/soundfont.sf2 -35   # out of bounds, reset to default.

The user is informed when a soundfont is loaded (or not loaded), along
with any attenuation or amplification percentage.