Prevent FluidSynth start when user selecter auto
On ALSA, MIDI port autodetection almost always fails, because it uses
some hardcoded ports instead of really looking for available clients.

In result, when user has soundfont selected on Linux, it will *always*
result in FS starting (unlike other OSes, which will pick their built-in
backends). This is bad, because it DOSBox mixer has no idea if anything
is playing, so we get steady stream of callbacks, thus wasting a lot of
CPU on synthesising silence and it noticably slows down the startup.

This change makes FluidSynth opt-in, it won't be selected automatically
any more.