Print info about configure flags for optional packages
Opusfile and FluidSynth are optional, recommended packages - we want
packagers to enable those features, but if they can't (because packages
are simply missing), we prefer they disable it instead of not packaging
dosbox-staging.

Given choice of making yet another "note to packagers" (which they won't
read) and placing info line in configure script (which can't be placed
after PKG_CHECK_MODULES failed because of mysterious autoconf reasons),
I select the latter.