Sync Nuked-OPL3 with upstream repo
After Nuked OPL3 emulator was merged to DOSBox Staging code, author
decided to place it in a dedicated GitHub repo:

https://github.com/nukeykt/Nuked-OPL3

Afterwards, he graciously accepted my contributions that brought the
code to state when we can be 100% in sync with his repo.

Rename files to the same names as upstream and sync state of to
upstream version ed1104, which includes following changes:

- Replace Bit* typedefs with C99 fixed-size types
- Use C linkage to make header usable in C++ projects

And make appropriate buildsystem changes to use these files.