Work around MSYS2/MinGW build issues
Most of these problems are results of MinGW not being able to decide if
it is POSIX or ISO compliant or not.

Replace non-POSIX but ISO-compliant _mkdir(const char*) function (which
works in MSVC, but does not work in MinGW) with non-POSIX and non-ISO
mkdir(const char*) function (because it works in MinGW, but only
generates a warning in MSVC).

Add config.h before any use of mkdir, as that's where we disable MSVC
warnings urging us to use ISO versions of POSIX functions (with '_'
prefix).

Use ISO _S_IFDIR, _S_IFMT and include sys/stat.h as otherwise MinGW
can't find POSIX S_IFDIR, S_IFMT.

Change ifdef to use C11, non-POSIX strerror_s in MinGW, because
GNU-provided, POSIX-like strerror_r is not available (which is weird,
because C99 stuff generally does not work in MinGW so why C11 stuff
works suddenly?).

Fixes: #753