Prefer MT-32 upstream configuration of "partials" limit
Although we could include a constexpr for the number of partials,
we instead defer to upstream to configure this correctly based
on their testing and knowledge of the hardware.

The Munt library and its external application synthesizers default to 32:
 - mt32emu/src/globals.h: MT32EMU_DEFAULT_MAX_PARTIALS 32
 - mt32emu/src/globals.h: DEFAULT_MAX_PARTIALS = MT32EMU_DEFAULT_MAX_PARTIALS;
 - mt32emu_qt/src/Master.cpp: synthProfile.partialCount .. default to DEFAULT_MAX_PARTIALS

If the time comes when we need to update this either for performance
reasons or to fix playback, then we can re-add it as a constexpr.